requirements.txt
aiohttp
python-rapidjson
certifi
pydantic[ujson]
